requirements.txt
python-dotenv>=1.0.1
langchain-aws>=0.2.5
SQLAlchemy>=2.0.32
psycopg[binary,pool]>=3.2.1
mcp>=1.2.0

